No longer explicitely load Dates extension, which is now configurable
